Extract validation functions from the app modules
Slowly, slowly trimming the flask app modules down to be just
about serving requests.
